+++ lib/common.sh
@@ -39,7 +39,7 @@ if [[ -f "${AI_WARP_SHARE}/VERSION" ]]; then
 elif [[ -f "${AI_WARP_SHARE_LIB:-}/VERSION" ]]; then
   AI_WARP_VERSION="$(tr -d '[:space:]' <"${AI_WARP_SHARE_LIB}/VERSION")"
 else
-  AI_WARP_VERSION="${AI_WARP_VERSION_PIN:-1.2.1}"
+  AI_WARP_VERSION="${AI_WARP_VERSION_PIN:-1.2.2}"
 fi
 readonly AI_WARP_VERSION